[004] 7zip plugin glue on UTF-8 (T051)
The bundled 7za carries an OPENSAL patch that disables USE_UNICODE_FSTRING, so
its file layer converts paths through the ACP - the glue therefore opens files
itself: new standalone IInStream/IOutStream implementations over CreateFileW +
extended-length paths (retry/abort semantics and the required QI surface kept).

Entry names now cross the interface as UTF-8 in both directions (listing,
extraction, packing, overwrite dialogs, progress); passwords deliberately stay
ACP (they come from the plugin's own ANSI dialogs). Interface path buffers moved
to the heap; a real unbounded lstrcpy overflow in SetArchiveName fixed.
Remaining ACP paths (plugin dialogs, 7za.dll module lookup) documented.

+++ src/plugins/zip/common2.cpp
@@ -260,7 +260,7 @@ void CZipCommon::FindLastFile(char* lastFile)
     pathLen = lstrlen(buf);
     do
     {
-        if (!FindDataNameU8(&data, foundName, sizeof(foundName)))
+        if (!FindDataNameU8(&data, foundName, (int)sizeof(foundName)))
             continue;
         SplitPath2(foundName, path, name, ext);
         /*if (!*name)
